refactor: extract logic for TodoInfo component
