+++ src/App.tsx
@@ -24,6 +24,6 @@ export const App: React.FC = () => (
   <div className="App">
     <h1 className="App__title">Static list of todos</h1>
 
-    <TodoList todoItems={todos} />
+    <TodoList todoListItems={todos} />
   </div>
 );
